Sign in with new user (testUser) composable working
