+++ apps/web/src/composables/router.ts
@@ -1,5 +1,4 @@
 import { createWebHistory, createRouter } from 'vue-router'
-import useUsers from '@/composables/users'
 import Overview from '@/pages/overview/Overview.vue'
 import Operator from '@/pages/operators/Operator.vue'
 
@@ -29,24 +28,6 @@ const router = createRouter({
     // eslint-disable-next-line @typescript-eslint/ban-ts-comment
     // @ts-ignore
     routes
-})
-
-// TO DO: Add a routing beforeEach that 
-// dynamically fixes rerouting to auth page
-
-router.beforeEach(async (to, from, next) => {
-    const { checkUserSessionExists } = useUsers()
-    await checkUserSessionExists()
-    next()
-    // if (to.fullPath === '/auth' && !loggedIn) {
-    //     next()
-    // } else if (to.fullPath === '/auth' && loggedIn) {
-    //     next('/')
-    // } else if (!loggedIn) {
-    //     next('/auth')
-    // } else {
-    //     next()
-    // }
 })
 
 export default router
