+++ apps/web/src/layouts/default-layout.vue
@@ -6,31 +6,87 @@ import router from '@/composables/router'
 import VueFeather from 'vue-feather'
 import useFormat from '@/composables/format'
 import useScreenDimensions from '@/composables/screenDimensions'
-import useUser from '@/composables/users'
-import useWallet from '@/composables/wallet'
+import useTestUser from '@/composables/testUser'
+import { CryptoAddress, Currency, ProviderString } from '@casimir/types'
+import useEthers from '@/composables/ethers'
+import useLedger from '@/composables/ledger'
+import useTrezor from '@/composables/trezor'
+import useWalletConnect from '@/composables/walletConnectV2'
 
+const { ethersProviderList, getEthersAddressWithBalance } = useEthers()
+const { convertString, trimAndLowercaseAddress } = useFormat()
+const { getLedgerAddress } = useLedger()
+const { getTrezorAddress } = useTrezor()
+const { screenWidth } = useScreenDimensions()
+const { connectWalletConnectV2 } = useWalletConnect()
+
+const activeWallets = [
+  'MetaMask',
+  'CoinbaseWallet',
+  'WalletConnect',
+  'Trezor',
+  'Ledger',
+  'TrustWallet',
+  // 'IoPay',
+] as ProviderString[]
 const authFlowCardNumber = ref(1)
-const selectedProvider = ref(null as null | string)
-const termsCheckbox = ref(true)
-
+const selectedProvider = ref(null as ProviderString | null)
 const openRouterMenu = ref(false)
-
-const { convertString } = useFormat()
+const openWalletsModal = ref(false)
+const walletProviderAddresses = ref([] as CryptoAddress[])
 
 const {
-  screenWidth
-} = useScreenDimensions()
-
-const  {
-  activeWallets,
-  openWalletsModal,
-  walletProviderAddresses,
+  user,
+  addAccountToUser,
+  login,
   logout,
-  selectAddress,
-  selectProvider,
-} = useWallet()
+} = useTestUser()
 
-const { user } = useUser()
+function checkIfAddressIsUsed (account: CryptoAddress): boolean {
+  const { address } = account
+  if (user.value?.accounts) {
+    const accountAddresses = user.value.accounts.map((account: any) => account.address)
+    if (accountAddresses.includes(address)) return true
+  }
+  return false
+}
+
+  /**
+   * Checks if user is adding an account or logging in
+   * @param address 
+  */
+   async function selectAddress(address: string) {
+    address = trimAndLowercaseAddress(address)
+    if (user.value) {
+      // Add account
+      await addAccountToUser({provider: selectedProvider.value as ProviderString, address, currency: 'ETH'})
+    } else {
+      // Login
+      await login({provider: selectedProvider.value as ProviderString, address, currency: 'ETH'})
+    }
+  }
+
+/**
+ * Sets the selected provider and returns the set of addresses available for the selected provider
+ * @param provider 
+ * @param currency 
+*/
+async function selectProvider(provider: ProviderString, currency: Currency = 'ETH'): Promise<void> {
+  console.clear()
+  try {
+    if (provider === 'WalletConnect') {
+      walletProviderAddresses.value = await connectWalletConnectV2('eip155:5') as CryptoAddress[]
+    } else if (ethersProviderList.includes(provider)) {
+      walletProviderAddresses.value = await getEthersAddressWithBalance(provider) as CryptoAddress[]
+    } else if (provider === 'Ledger') {
+      walletProviderAddresses.value = await getLedgerAddress[currency]() as CryptoAddress[]
+    } else if (provider === 'Trezor') {
+      walletProviderAddresses.value = await getTrezorAddress[currency]() as CryptoAddress[]
+    }
+  } catch (error: any) {
+    throw new Error(`Error selecting provider: ${error.message}`)
+  }
+}
 
 const show_setting_modal = ref(false)
 
@@ -176,6 +232,7 @@ onUnmounted(() =>{
 
           <div class="connect_wallet_gradient">
             <button
+              id="connect_wallet_button"
               class="connect_wallet flex justify-between items-center gap-[8px] whitespace-nowrap"
               @click="openWalletsModal = true"
             >
@@ -327,29 +384,25 @@ onUnmounted(() =>{
                 </h6>
               </div>
               <div v-else>
-                <!-- Not needed since its the same on the staking component -->
-                <!-- <div class="my-[20px] px-[10px] nav_items">
-                  <input
-                    v-model="termsCheckbox"
-                    type="checkbox"
-                    class="mr-[5px]"
-                  > By connecting my address, I certify that I have read and accept the updated 
-                  <button class="text-primary hover:text-blue_3">
-                    Terms of Service
-                  </button>.
-                </div>  -->
                 <button
                   v-for="act in walletProviderAddresses"
                   :key="act.address"
                   class="w-full border rounded-[8px] px-[10px] py-[15px] flex flex-wrap gap-[10px] text-center items-center justify-between hover:border-blue_3 mb-[10px]"
+                  :disable="checkIfAddressIsUsed(act)"
                   @click="selectAddress(act.address), openWalletsModal = false, authFlowCardNumber = 1"
                 >
                   <div>
-                    {{ convertString(act.address) }}
+                    {{ convertString(act.address) }} 
                   </div>
                   <div>
                     {{ parseFloat(parseFloat(act.balance).toFixed(2)) }} ETH
                   </div>
+                  <p
+                    v-if="checkIfAddressIsUsed(act)"
+                    class="text-decline text-[12px] font-[400]"
+                  >
+                    Address in use.
+                  </p>
                 </button>
               </div>
             </div>
