+++ common/types/src/interfaces/LoginCredentials.ts
@@ -5,6 +5,4 @@ export interface LoginCredentials {
     address: string
     provider: ProviderString
     currency: Currency
-    message: string
-    signedMessage: string
 }
