+++ common/types/src/interfaces/SignInWithEthereumCredentials.ts
@@ -1,0 +1,6 @@
+import { LoginCredentials } from './LoginCredentials'
+
+export interface SignInWithEthereumCredentials extends LoginCredentials {
+    message: string
+    signedMessage: string
+}
